Added user community validation to TaxesController.

+++ Source/Server/NeighboursCommunitySystem.API/Controllers/TaxesController.cs
@@ -7,21 +7,29 @@
     using AutoMapper;
     using AutoMapper.QueryableExtensions;
     using System;
+    using Microsoft.AspNet.Identity;
+    using Models;
+    using Data.Repositories;
+    using Data.DbContexts;
 
     public class TaxesController : ApiController
     {
         private readonly ITaxesService taxes;
+        private readonly ICommunitiesService communities;
+        private readonly string currentUserId;
 
-        public TaxesController(ITaxesService taxes)
+        public TaxesController(ITaxesService taxes, ICommunitiesService communities)
         {
             this.taxes = taxes;
+            this.communities = communities;
+            this.currentUserId = this.User.Identity.GetUserId();
         }
 
         [Authorize(Roles = "DbAdmin")]
         public IHttpActionResult Get()
         {
             var allTaxes = taxes.All().ProjectTo<TaxDataTransferModel>().ToList();
-
+            
             return this.Ok(allTaxes);
         }
 
@@ -37,10 +45,15 @@
         [Authorize(Roles = "DbAdmin,Administrator,Accountant")]
         public IHttpActionResult Community(int id)
         {
+            if (!ValidateCurrentUserCommunity(communities.GetById(id)))
+            {
+                return this.Unauthorized();
+            }
+
             var communityTaxes = taxes
-                                .GetByCommunityId(id)
-                                .ProjectTo<TaxDataTransferModel>()
-                                .ToList();
+                                    .GetByCommunityId(id)
+                                    .ProjectTo<TaxDataTransferModel>()
+                                    .ToList();
 
             return this.Ok(communityTaxes);
         }
@@ -48,6 +61,16 @@
         [Authorize(Roles = "DbAdmin,Administrator")]
         public IHttpActionResult Post(TaxRequestTransferModel model)
         {
+            if (!ValidateCurrentUserCommunity(communities.GetById(model.CommunityId)))
+            {
+                return this.Unauthorized();
+            }
+
+            if (!this.ModelState.IsValid)
+            {
+                return this.BadRequest(this.ModelState);
+            }
+
             int taxId = taxes.Add(model);
 
             return this.Ok(taxId);
@@ -56,6 +79,11 @@
         [Authorize(Roles = "DbAdmin,Administrator")]
         public IHttpActionResult Delete(int id)
         {
+            if (!ValidateCurrentUserCommunity(communities.GetById(taxes.GetById(id).CommunityId)))
+            {
+                return this.Unauthorized();
+            }
+
             taxes.DeleteById(id);
 
             return this.Ok();
@@ -64,6 +92,16 @@
         [Authorize(Roles = "DbAdmin,Administrator")]
         public IHttpActionResult Put(int id, TaxDataTransferModel model)
         {
+            if (!ValidateCurrentUserCommunity(communities.GetById(taxes.GetById(id).CommunityId)))
+            {
+                return this.Unauthorized();
+            }
+
+            if (!this.ModelState.IsValid)
+            {
+                return this.BadRequest(this.ModelState);
+            }
+
             taxes.UpdateById(id, model);
 
             return this.Ok();
@@ -73,6 +111,11 @@
         [Authorize(Roles = "DbAdmin,Administrator,Accountant")]
         public IHttpActionResult Available(int id)
         {
+            if (!ValidateCurrentUserCommunity(communities.GetById(id)))
+            {
+                return this.Unauthorized();
+            }
+
             var availableTaxes = taxes
                                     .GetByCommunityId(id)
                                     .Where(t => t.Deadline > DateTime.Now)
@@ -86,13 +129,24 @@
         [Authorize(Roles = "DbAdmin,Administrator,Accountant")]
         public IHttpActionResult Expired(int id)
         {
+            if (!ValidateCurrentUserCommunity(communities.GetById(id)))
+            {
+                return this.Unauthorized();
+            }
+
             var expiredTaxes = taxes
-                                    .GetByCommunityId(id)
-                                    .Where(t => t.Deadline < DateTime.Now)
-                                    .ProjectTo<TaxDataTransferModel>()
-                                    .ToList();
+                                .GetByCommunityId(id)
+                                .Where(t => t.Deadline < DateTime.Now)
+                                .ProjectTo<TaxDataTransferModel>()
+                                .ToList();
 
             return this.Ok(expiredTaxes);
+        }
+
+        [NonAction]
+        public bool ValidateCurrentUserCommunity(Community currentCommunity)
+        {
+            return currentCommunity.Users.Any(u => u.Id == currentUserId);
         }
     }
 }
